requirements.txt
requests[socks]
beautifulsoup4
fpdf2
transformers
scikit-learn
torch
Flask 